refactor: Modularizing static files and refactoring an assignValues function
