+++ src/components/Grid/Grid.jsx
@@ -2,7 +2,7 @@
 import { useContext, useEffect, useState } from 'react'
 import { GridContext } from '../../context/GridContext.jsx'
 import { getRandomNumber } from '../../utils/findCell.js'
-import alphabet from '../../utils/test-board-1.json'
+import alphabet from '../../assets/files/test-board-1.json'
 import Card from '../Card/Card.jsx'
 import './Grid.css'
