+++ src/components/Word/Word.jsx
@@ -1,7 +1,7 @@
 /* eslint-disable react-hooks/exhaustive-deps */
 import { useContext, useEffect } from 'react'
 import { GridContext } from '../../context/GridContext'
-import dictionary from '../../utils/dictionary.json'
+import dictionary from '../../assets/files/dictionary.json'
 import './Word.css'
 
 const Word = () => {
